+++ build.gradle
@@ -1,21 +1,13 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-	ext {
-		// Testing
-		assertJVersion = '1.7.1'
-		expressoVersion = '2.2.2'
-		jUnitVersion = '4.12'
-		robolectricVersion = '3.5.1'
-		testingSupportLibVersion = '0.1'
-	}
 	repositories {
 		jcenter()
 		mavenCentral()
 		google()
 	}
 	dependencies {
-		classpath 'com.android.tools.build:gradle:4.0.1'
-		classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.4.10"
+		classpath 'com.android.tools.build:gradle:4.2.0-beta01'
+		classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.4.20'
 	}
 }
