+++ tireguide/src/main/java/com/droidsmith/tireguide/calcengine/Calculator.kt
@@ -1,6 +1,6 @@
 package com.droidsmith.tireguide.calcengine
 
-import java.util.Objects
+import java.util.*
 
 /**
  * Calculates the proper tire pressure for a bicycle tire based on weight and width of tire.
