+++ tireguide/src/main/java/com/droidsmith/tireguide/splash/SplashActivity.kt
